jpg_to_png_converter/requirements.txt
autopep8==2.0.2
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
idna==3.4
libretranslatepy==2.1.1
lxml==4.9.2
oauthlib==3.2.2
Pillow==9.5.0
pycodestyle==2.10.0
pygame==2.4.0
pyjokes==0.6.0
PyPDF2==3.0.1
requests==2.28.2
requests-oauthlib==1.3.1
soupsieve==2.4.1
translate==3.6.1
tweepy==4.14.0
urllib3==1.26.15
